refactor(VmTurnZone): use the browser microtask queue for JS
