+++ tools/broccoli/html-replace/SCRIPTS.html
@@ -1,5 +1,4 @@
-<script src="es6-promise.js" type="text/javascript"></script>
-<script src="zone.js" type="text/javascript"></script>
+<script src="zone-microtask.js" type="text/javascript"></script>
 <script src="long-stack-trace-zone.js" type="text/javascript"></script>
 <script src="traceur-runtime.js" type="text/javascript"></script>
 <script src="es6-module-loader-sans-promises.src.js" type="text/javascript"></script>
