+++ tools/broccoli/html-replace/SCRIPTS_benchmarks.html
@@ -1,5 +1,4 @@
-<script src="es6-promise.js" type="text/javascript"></script>
-<script src="zone.js" type="text/javascript"></script>
+<script src="zone-microtask.js" type="text/javascript"></script>
 <script src="long-stack-trace-zone.js" type="text/javascript"></script>
 <script src="url_params_to_form.js" type="text/javascript"></script>
 <script src="traceur-runtime.js" type="text/javascript"></script>
