+++ tools/broccoli/trees/browser_tree.ts
@@ -78,8 +78,7 @@ module.exports = function makeBrowserTree(options, destinationPath) {
 
   var vendorScriptsTree = flatten(new Funnel('.', {
     files: [
-      'zone/es6-promise.js',
-      'zone/zone.js',
+      'zone/zone-microtask.js',
       'zone/long-stack-trace-zone.js',
       'node_modules/es6-module-loader/dist/es6-module-loader-sans-promises.src.js',
       'node_modules/systemjs/dist/system.src.js',
